이미지, 파일 체크 시 삭제되도록 로직 수정

+++ src/main/java/dhi/ds/controllers/SolutionController.java
@@ -226,7 +226,7 @@ public class SolutionController {
             List<Integer> chList = fileCheckList.get();
             for(int i =0; i < chList.size(); i++){
                 for(int j=0; j < fileIdList.size(); j++) {
-                    if(fileIdList.get(j) == chList.get(i)) {
+                    if(fileIdList.get(j).intValue() == chList.get(i).intValue()) {
                         fileIdList.remove(j);
                         break;
                     }
